ArmPlatformPkg/ArmVExpressPkg: enable UEFI Secure Boot
This allows the FVP target to be built with UEFI Secure Boot enabled,
by passing -D SECURE_BOOT_ENABLE to the build command line. Note that
you will need to disable the ARM BDS (-D USE_ARM_BDS=FALSE) or you will
not be able to enroll certificates, since the ARM BDS does not provide
a GUI to do so.

The FVP Fast model is recommended in this case, since the certificate
store is kept in NOR flash.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>