Platforms/ARM: TC2: Add LAN9118 PCD
The LAN9118 driver in EDK2 has been updated to include a PCD for
over-riding the default auto-negotiation timeout.

The default setting of 4000us is not long enough for TC2, so set it to
400000.

NOTE: LAN9118 on TC2 is only working in RELEASE builds at the moment.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>