Platforms/ARM: add PcdUartDefaultReceiveFifoDepth
The upstream code uses a PCD to set the FIFO depth:

f423d76  2016-03-17  MdeModulePkg/SerialDxe: Set FIFO depth with PCD

Add a sensible default for the PCD before the code becomes more widely
used.

The PL011UartInitializePort function in the PL011 driver expects the
receive FIFO depth to be set to a sane value or zero to use the
appropriate default for the version of PL011 present on the device.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>