Platforms/Juno: drop FACS table
The FACS table does not expose anything meaningful on AArch64, and since
the Tianocore ACPI table handling code insists on locating FACS below 4 GB,
which may fail since since AArch64 platforms may not have any system RAM
below 4 GB to begin with, having its definition in the source tree does
more harm than good.

The reason for this behavior is to ensure that a 32-bit PEI can access the
FACS table on an otherwise 64-bit system, but this is a concern that does not
apply to AArch64, since PEI always runs in 64-bit mode in that case. The PI
spec currently does not provide any means for PEI to convey its bitness or
how much system RAM it can access, so a permanent fix requires a spec update
first.

So simply remove the FACS table until the PIWG clarifies the spec in this
regard.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Ryan Harkin <ryan.harkin@linaro.org>
Reviewed-by: Graeme Gregory <graeme.gregory@linaro.org>