Platforms/ARM/FVP: convert to alsc to improve build reliability
.asl files can end up being modified to later versions of spec when FADT
is tied to an earlier version. This can cause builds to break
unexpectedly on tools upgrades. Convert to aslc to fix the version of
ACPI in use for this platform to 6.1

Note that this version of the MADT table describes the GIC as a v3,
which is the only version we support for the Foundation model at the
moment (given that its GICv2 lives at a different address than its
FVP Base counterpart)

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Graeme Gregory <graeme.gregory@linaro.org>
[ardb: mention GICv3]
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>