Platforms/Styx: add FvNameGuid to FvMain FV
This adds name GUIDs to the Styx FDFs so that DevicePath references to
its contents (i.e., for the UiApp and Shell UEFI applications) are
emitted unambiguously rather than as MemoryMapped()/ device path nodes,
whose values may change between boots (e.g., if the amount of memory
changes, or if the firmware image is updated).

Having stable boot entries is preferable, given that the generic BDS code
autogenerates entries for the UI app and the shell, but does not clean up
any stale outdated ones that refer to FV files that no longer exist.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>