Hisilicon/D0x: get rid of APRIORI declarations
APRIORI declarations completely bypass the ordinary dependency based
dynamic dispatch of PEIMs and DXE_DRIVERs, and so should be avoided if
possible. The reason Pcd.inf requires it in this case is because it has
a circular dependency on itself due to its dependency on PcdLib, which
resolves to a library class resolution that introduces a dependency
on a protocol Pcd.inf produces itself (or a PPI in case of the PEIM version).

So override the PcdLib resolution for Pcd.inf to the Null PcdLib version,
which does not introduce this protocol dependency. This allows us to
simply remove the APRIOR PEI|DXE declarations.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Heyi Guo <heyi.guo@linaro.org>