Drivers/Xhci: implement firmware upload driver for Renesas PD72020x
This implements a PCI driver for the Renesas PD720201/PD720202 PCIe
XHCI controller, which in some cases requires a device firmware to be
uploaded by the system firmware in order to function.

By registering a UEFI driver with high dispatch priority, this driver
is able to access the device early, install the firmware, and then decline
the driver dispatch so that the device is fully functional when the real
XHCI driver gets to see it.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>