D02/D03: workaround to fix timer interrupt issue
Timer interrupt cannot be reported after EDK2 commit 7989300, for GIC
on D02/D03 is not fully ARM GIC compliant. The issue has been fixed on
newer chips so we use a workaround for D02 and D03 only.
On D02 and D03, IRQ will be latched in GIC logic except virtual timer
interrupt IRQ #27, so we change to use virtual timer instead of physical
in UEFI.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Heyi Guo <heyi.guo@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>