Platforms/AMD/Styx: Add configuration for Overdrive 1000
Add a configuration for the SoftIron Overdrive 1000. This is similar to
OverdriveBoard/ configuration (which is for the Overdrive 3000).

The Overdrive 1000 contains an AMD Opteron-A-based SoC with 4
Cortex-A57 cores. The board contains Marvell Yukon 1Gbit ethernet, two
SATA 3.0 ports, and two USB 3.0 ports.

This configuration contains a PreUefiFirmware.bin image which was
generated using the AMD Firmware Development Kit (FDK) version 1.0.0.2,
configured for use with DDR4 memory.

In comparison with the .dsc and .fdf in OverdriveBoard/, the files for
Overdrive1000Board contain the following differences:
 * Number of cores,
 * XGBE support removed (non-existent Overdrive 1000),
 * SataPi and SataPortCount for available ports,
 * PCIe configuration,
 * Addition of MarvellYukonDxe driver,
 * Switch to MdePkg's BaseMemoryLib and BaseMemoryLibOptPxe,
 * DTS strings,
 * New GUID,
 * Remove RVCT Flags,
 * Remove disabling of diagnostic flags,

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Alan Ott <alan@softiron.co.uk>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>