Platform/Marvell: Extend and share the platform description
Move the platform description structure to a common header
HwDescLib.h. Extend the structure with entries corresponding
to interfaces located on a first CP slave device, so that both
CP110 blocks are described. Change the platform description
structure type to MVHW_NONDISCOVERABLE_DESC. Also add macros
for array sizes and base addresses.

This change is a beginning of Armada 7k/8k support improvement,
that will allow to store SoC's description in one place, accessible
for the drivers and libraries. With that, the PCD description will be
cleaner and it will also ease adding support for multiple interfaces
of one kind.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>