Platforms/StyxSpiFvDxe: fix potential boot crash on varstore write
The varstore shadow FV is kept in sync with actual SPI flash read,
write and erase operations. Since we only expose a small slice of
the SPI flash for the variable store, we keep an internal LBA offset
and take it into account when translating shadow FV LBAs to actual
LBAs.

As it turns out, the erase routine applies the LBA offset incorrectly,
resulting in the wrong flash block being erased, and the wrong range
to be erased in the shadow FV, which could result in a crash if the
memory access is out of bounds.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>