Marvell/Armada7k8k: Add PciSegmentLib implementation
Implement a special version of PciSegmentLib that takes the quirky
nature of the Synopsys Designware PCIe IP into account. In particular,
we need to ignore config space accesses to all devices on the first
bus except device 0, because the broadcast nature of type 0 configuration
cycles will result in whatever device is in the slot to appear at each
of the 32 device positions. The patch is based on Socionext Synquacer
PciSegmentLib implementation and will later be extended to support
multiple PCIE slots.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>