Marvell/Armada80x0McBin: add ACPI description of the PCI root
Add a MCFG table and add a description to the DSDT of the PCIe root
complex on the 80x0 McBin board. Note that the Synopsys IP is not
fully compliant, and requires a workaround in order to support ACPI
boot which reduces the bus range to a single bus.

Change-Id: I551e423b21bbc610d5a7713e419858331d20f787
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52042
Tested-by: iSoC Platform CI <ykjenk@marvell.com>