Marvell/Applications: MdioCmd: Introduce mdio command
Marvell Armada SoC has SMI bus to control the PHY/Switch
that connects to it.
This patch introduces application of mdio command
to configure (read/write) PHY/Switch registers, with
protocol Mdio.
The command takes input Bus number, Address and Register
number, to read or write its value.

There is a dedicated command to read counters for Marvell
SOHO switch, which is by default disabled, and should only
be enabled for the boards that use it.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: jinghua <jinghua@marvell.com>
Reviewed-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>