Platforms/Marvell/Armada: Add PP2 description in McBin Dsdt table
This commit introduces ACPI description for PP2 controllers on
Armada-8040-McBin board, which is handled in Linux by mvpp2 driver.
A new Hardware ID for the controller is added: "MRVL0110"

Until the PHY handling for ACPI is developed in Linux, port status
can be handled only with usage of the link interrupts.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>