Marvell/Drivers: MvSpiDxe: Enable using driver in RT
This patch applies necessary modifications, which allow to use
MvSpiDxe driver in variable support as a runtime service.
Its type is modified to DXE_RUNTIME_DRIVER, as well as
a new callback is introduced as a part of the SpiMasterProtocol.
It is supposed to configure the memory space for mmio access to
the host controller registers. Moreover gBS locking usage in
MvSpiTransfer is limited to the firmware, as the runtime access
to the flash is protected within the OS.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>