Marvell/Library: UtmiPhyLib: Switch to use MARVELL_BOARD_DESC protocol
UTMI driver used to get Armada7k8k UTMI controller description from
hardcoded values stored in the header file MvHwDescLib.h.
As a result it is very hard to support other Armada SoC families
with this driver.

This patch updates the driver to get UTMI controller description from
newly introduced MARVELL_BOARD_DESC protocol, and removes the dependency
on the hardcoded structures. Use PlatInitDxe driver to open the
protocol and pass information to library init. It will be used
for the ComPhyLib and MppLib as well.

Change-Id: I9f2e76a6d367dc07b2cf50254c9867dd9a22b795
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52604
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>