Marvell/Applications: Extend functionality of fupdate to flash on SD
Extend fupdate shell application to be able to flash firmware
on SD and MMC cards. Device on which firmware will be updated
is selected via third argument of command. When no device
is selected spi is flashed.
Command uses BlockIo protocol to communicate with SD/MMC.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Tomasz Michalec <tm@semihalf.com>