Marvell/Armada37x0Db: Shift main FX from 0x0 address
When using PEI phase, UEFI interprets 0x0 address
of boot FV as an error. In order to avoid it, it is shifted
to 0x1000 and to meet a hardcoded 'jump to 0x1000' instruction
at offset 0x0. This patch is a preparation for using PEI
by Armada platforms.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Jaroslaw Bieniek <shino@semihalf.com>