Marvell/Drivers: MvFvbDxe: Introduce non-mmio mode
This path enables support for reading variables directly from flash without
relying on it to be memory mapped. It adds PcdSpiMemoryMapped PCD that
allows to switch between the modes. When in non-memory-mapped mode the
driver will copy the variables from flash to previously allocated buffer
and set PcdFlashNvStorageVariableBase64, PcdFlashNvStorageFtwWorkingBase64
and PcdFlashNvStorageFtwSpareBase64 accordingly. Because of that satisfy
a depex dependency with the FTW driver by hooking NvVarStoreFormattedLib.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>