Marvell/Library: Introduce MvGpioLib class
Following patches will add two drivers that are capable
of producing the generic EMBEDDED_GPIO protocol. In order
to handle a situation, when the consumer module attempts
to locate and open a proper protocol, add a library
helper function.

In order to ease description of used GPIO pins and controllers
of the Armada 7k8k platforms, add the common enum types.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>