Marvell/Drivers: XenonDxe: Explicitly disable HS400
On another SoC revision, the capability register marks HS400 support
as enabled. However in case the interface itself is powered with 3.3V
this flag must be unset by the SdMmcOverride protocol callback -
otherwise the generic EmmcSwitchToHS400 () would be executed
with a failure.

Ensure that in case of SlowMode or 3.3V operation, the HS400 capability
will be disabled in the SdMmc driver, along with other highest-speed
modes.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>